ProgressBar class and move types into panel_config.ts
