+++ src/mapper.ts
@@ -1,242 +1,12 @@
-import { PanelConfig } from './panel_config';
-import { getFormattedValue } from './value_formatter';
+import { ProgressBar } from './progress_bar';
+import { PanelConfig, StatType } from './panel_config';
 
-import { TitleViewOptions } from './module';
 
 import * as _ from 'lodash';
 
 
 type KeyValue = [string, number];
 
-export enum StatType {
-  CURRENT = 'current',
-  MIN = 'min',
-  MAX = 'max',
-  TOTAL = 'total'
-};
-
-type TitleViewParams = {
-  barHeight: number,
-  titleTopMargin: number,
-  valueTopMargin: number
-};
-
-export class MultiProgressItem {
-  constructor(
-    private _panelConfig: PanelConfig,
-    private _title: string,
-    private _keys: string[],
-    private _values: number[],
-    private _maxValue: number
-  ) {
-    if(this._keys.length !== this._values.length) {
-      throw new Error('keys amount should be equal to values amount');
-    }
-  }
-
-  get title(): string {
-    return this._title;
-  }
-
-  get keys(): string[] {
-    return this._keys;
-  }
-
-  get values(): number[] {
-    return this._values;
-  }
-
-  get sumOfValues(): number {
-    return _.sum(this.values);
-  }
-
-  get percentValues(): number[] {
-    // TODO: this.sumOfValues * 1.1 is a hack to make sure bars don't wrap
-    // (they are wrapped when total width > 98%)
-    return this.values.map(
-      value => Math.floor(value / (this.sumOfValues * 1.1) * 100)
-    );
-  }
-
-  get colors(): string[] {
-    // TODO: customize colors
-    return ['green', 'yellow', 'red'];
-  }
-
-  get aggregatedProgress(): number {
-    return (_.sum(this.values) / this._maxValue) * 100;
-  }
-
-  get formattedValue(): string {
-    return getFormattedValue(
-      this.sumOfValues,
-      this._panelConfig.getValue('prefix'),
-      this._panelConfig.getValue('postfix'),
-      this._panelConfig.getValue('decimals')
-    );
-  }
-
-  get titleParams(): TitleViewParams {
-    const titleType = this._panelConfig.getValue('titleViewType');
-
-    switch(titleType) {
-      case TitleViewOptions.SEPARATE_TITLE_LINE:
-        return {
-          barHeight: 8,
-          titleTopMargin: 0,
-          valueTopMargin: -12
-        };
-      case TitleViewOptions.INLINE:
-        return {
-          barHeight: 20,
-          titleTopMargin: -20,
-          valueTopMargin: -18
-        };
-      default:
-        throw new Error(`Wrong titleType: ${titleType}`);
-    }
-  }
-
-  get opacity(): string {
-    return this._panelConfig.getValue('opacity');
-  }
-}
-
-export class ProgressItem {
-
-  private _panelConfig: PanelConfig;
-
-  private _key: string;
-  private _aggregatedValue: number;
-  private _maxAggregatedValue: number;
-  private _currentValue: number;
-  private _currentMaxValue: number;
-
-  constructor(
-    panelConfig: PanelConfig,
-    key: string,
-    aggregatedValue: number,
-    maxAggregatedValue: number,
-    currentValue: number,
-    currentMaxValue: number
-  ) {
-    this._panelConfig = panelConfig;
-    this._key = key;
-    this._aggregatedValue = aggregatedValue;
-    this._maxAggregatedValue = maxAggregatedValue;
-    // TODO: currentValue and currentMaxValue is not the best idea
-    this._currentValue = currentValue;
-    this._currentMaxValue = currentMaxValue;
-  }
-
-  get title(): string {
-    return this._key;
-  }
-
-  get aggregatedProgress(): number {
-    return 100 * this._aggregatedValue / this._maxAggregatedValue;
-  }
-
-  get currentProgress(): number {
-    return 100 * this._currentValue / this._currentMaxValue;
-  }
-
-  get maxValue(): number {
-    return this._maxAggregatedValue;
-  }
-
-  get aggregatedValue(): number {
-    return this._aggregatedValue;
-  }
-
-  get currentValue(): number {
-    return this._currentValue;
-  }
-
-  get currentFormattedValue(): string {
-    const value =
-      this._panelConfig.getValue('valueLabelType') === 'percentage' ?
-        this.currentProgress :
-        this._currentValue;
-    return getFormattedValue(
-      value,
-      this._panelConfig.getValue('prefix'),
-      this._panelConfig.getValue('postfix'),
-      this._panelConfig.getValue('decimals')
-    );
-  }
-
-  get formattedValue(): string {
-    const value =
-      this._panelConfig.getValue('valueLabelType') === 'percentage'
-        ? this.aggregatedProgress
-        : this._aggregatedValue;
-    return getFormattedValue(
-      value,
-      this._panelConfig.getValue('prefix'),
-      this._panelConfig.getValue('postfix'),
-      this._panelConfig.getValue('decimals')
-    );
-  }
-
-  get opacity(): string {
-    return this._panelConfig.getValue('opacity');
-  }
-
-  get color(): string {
-    var colorType = this._panelConfig.getValue('coloringType');
-    if(colorType === 'auto') {
-      return 'auto'
-    }
-    if(colorType === 'thresholds') {
-      var thresholdsStr = this._panelConfig.getValue('thresholds');
-      var colors = this._panelConfig.getValue('colors');
-      var value = this.aggregatedProgress;
-      if(thresholdsStr === undefined) {
-        return colors[0];
-      }
-      var thresholds = thresholdsStr.split(',').map(parseFloat);
-      for(var i = thresholds.length; i > 0; i--) {
-        if(value >= thresholds[i - 1]) {
-          return colors[i];
-        }
-      }
-      return colors[0];
-    }
-    if(colorType === 'key mapping') {
-      var colorKeyMappings = this._panelConfig.getValue('colorKeyMappings') as any[];
-      var keyColorMapping = _.find(colorKeyMappings, k => k.key === this._key);
-      if(keyColorMapping === undefined) {
-        return this._panelConfig.getValue('colorsKeyMappingDefault');
-      }
-      return keyColorMapping.color;
-    }
-
-    throw new Error('Unknown color type ' + colorType);
-  }
-
-  get titleParams(): TitleViewParams {
-    const titleType = this._panelConfig.getValue('titleViewType');
-
-    switch(titleType) {
-      case TitleViewOptions.SEPARATE_TITLE_LINE:
-        return {
-          barHeight: 8,
-          titleTopMargin: 0,
-          valueTopMargin: -22
-        };
-      case TitleViewOptions.INLINE:
-        return {
-          barHeight: 20,
-          titleTopMargin: -20,
-          valueTopMargin: -18
-        };
-      default:
-        throw new Error(`Wrong titleType: ${titleType}`);
-    }
-  }
-
-}
 
 export class Mapper {
 
@@ -248,7 +18,7 @@ export class Mapper {
     this._templateSrv = templateSrv;
   }
 
-  mapMetricData(seriesList: any): ProgressItem[] {
+  mapMetricData(seriesList: any): ProgressBar[] {
     const statType: StatType = this._panelConfig.getValue('statNameOptionValue');
     const mappingType = this._panelConfig.getValue('mappingType');
     const statProgressType = this._panelConfig.getValue('statProgressType');
@@ -263,80 +33,44 @@ export class Mapper {
     let kstat: KeyValue[] = [];
     let currentStat: KeyValue[] = [];
 
-    if(seriesList[0].columns !== undefined) {
-      let keys = seriesList[0].columns.map(col => col.text);
+  
+    let keys = seriesList[0].columns.map(col => col.text);
 
-      const keyColumn = this._panelConfig.getValue('keyColumn');
-      let keyIndex = 0;
-      if(keyColumn !== '') {
-        keyIndex = keys.findIndex(key => key === keyColumn);
-      }
+    const keyColumn = this._panelConfig.getValue('keyColumn');
+    let keyIndex = 0;
+    if(keyColumn !== '') {
+      keyIndex = keys.findIndex(key => key === keyColumn);
+    }
 
-      const title = keys[keyIndex];
+    const title = keys[keyIndex];
 
-      let skipIndexes: number[] = [keyIndex];
-      const skipColumn = this._panelConfig.getValue('skipColumn');
-      if(skipColumn !== '') {
-        skipIndexes.push(keys.findIndex(key => key === skipColumn));
-      }
+    let skipIndexes: number[] = [keyIndex];
+    const skipColumn = this._panelConfig.getValue('skipColumn');
+    if(skipColumn !== '') {
+      skipIndexes.push(keys.findIndex(key => key === skipColumn));
+    }
 
-      const maxValue = _.max(
-        seriesList[0].rows.map(
-          row => _.sum(
-            row.filter((value, idx) => !_.includes(skipIndexes, idx))
-          )
+    const maxValue = _.max(
+      seriesList[0].rows.map(
+        row => _.sum(
+          row.filter((value, idx) => !_.includes(skipIndexes, idx))
         )
-      );
-
-      const filteredKeys = keys.filter((key, idx) => !_.includes(skipIndexes, idx));
-      
-      // TODO: it's wrong, we return a bad type here
-      return seriesList[0].rows.map(
-        row => new MultiProgressItem(
-          this._panelConfig,
-          row[keyIndex],
-          filteredKeys,
-          row.filter((value, idx) => !_.includes(skipIndexes, idx)),
-          maxValue as number
-        )
-      );
-    }
-
-    if(mappingType === 'datapoint to datapoint') {
-      if(statType === StatType.TOTAL && seriesList.length == 1) {
-        kstat = this._mapKeysTotal(seriesList);
-      } else {
-        kstat = this._mapNumeric(seriesList, statType, nullMapping);
-      }
-      currentStat = this._mapNumeric(seriesList, StatType.CURRENT, nullMapping);
-    } else {
-      kstat = this._mapTargetToDatapoints(seriesList, statType);
-      currentStat = this._mapTargetToDatapoints(seriesList, StatType.CURRENT);
-    }
-
-    const maxValue = this._getMaxValue(kstat, statProgressType, statProgressMaxValue);
-    const currentMaxValue = this._getMaxValue(currentStat, statProgressType, statProgressMaxValue);
-
-    if(alias !== '') {
-      kstat.forEach(k => {
-        const scopedVars = {
-          __key: { value: k[0] }
-        };
-        k[0] = this._templateSrv.replace(alias, scopedVars);
-      });
-    }
-
-    return _.map(
-      kstat,
-      (k: KeyValue, idx: number) => new ProgressItem(
-        this._panelConfig,
-        k[0],
-        k[1],
-        maxValue,
-        currentStat[idx][1],
-        currentMaxValue
       )
     );
+
+    const filteredKeys = keys.filter((key, idx) => !_.includes(skipIndexes, idx));
+    
+    // TODO: it's wrong, we return a bad type here
+    return seriesList[0].rows.map(
+      row => new ProgressBar(
+        this._panelConfig,
+        row[keyIndex],
+        filteredKeys,
+        row.filter((value, idx) => !_.includes(skipIndexes, idx)),
+        maxValue as number
+      )
+    );
+    
   }
 
   // TODO: enum statProgressType
