+++ src/module.ts
@@ -1,16 +1,16 @@
-import { GraphTooltip, TooltipMode, TooltipItem } from './graph_tooltip';
-import { PanelConfig } from './panel_config';
-import { Mapper, ProgressItem, StatType } from './mapper';
+import { GraphTooltip, TooltipItem } from './graph_tooltip';
+import * as PanelConfig from './panel_config';
+import { Mapper } from './mapper';
+import { ProgressBar } from './progress_bar';
 
 import { MetricsPanelCtrl, loadPluginCss } from 'grafana/app/plugins/sdk';
 
 import * as _ from 'lodash';
 
-
-export enum TitleViewOptions {
-  SEPARATE_TITLE_LINE = 'Separate title line',
-  INLINE = 'Inline'
-};
+export type HoverEvent = {
+  index: number,
+  event: any
+}
 
 const ERROR_MAPPING = `
   Can't map the received metrics,
@@ -18,58 +18,30 @@ const ERROR_MAPPING = `
 `;
 const ERROR_NO_DATA = "no data";
 
-export type HoverEvent = {
-  index: number,
-  event: any
-}
-
-const DEFAULTS = {
-  keyColumn: '',
-  // TODO: skip multiple columns
-  skipColumn: '',
-  statNameOptionValue: StatType.CURRENT,
-  statProgressType: 'shared',
-  statProgressMaxValue: null,
-  coloringType: 'auto',
-  titleViewType: TitleViewOptions.SEPARATE_TITLE_LINE,
-  sortingOrder: 'none',
-  valueLabelType: 'percentage',
-  mappingType: 'datapoint to datapoint',
-  alias: '',
-  prefix: '',
-  postfix: '',
-  thresholds: '10, 30',
-  // https://github.com/grafana/grafana/blob/v4.1.1/public/app/plugins/panel/singlestat/module.ts#L57
-  colors: ['rgba(245, 54, 54, 0.9)', 'rgba(237, 129, 40, 0.89)', 'rgba(50, 172, 45, 0.97)'],
-  colorsKeyMappingDefault: 'rgba(245, 54, 54, 0.9)',
-  colorKeyMappings: [],
-  nullMapping: undefined,
-  tooltipMode: TooltipMode.ALL_SERIES,
-  opacity: 0.5
-};
-
 class Ctrl extends MetricsPanelCtrl {
   static templateUrl = 'partials/template.html';
 
   public mapper: Mapper;
-  public items: ProgressItem[];
 
-  private _panelConfig: PanelConfig;
+  // TODO: rename progressBars
+  public items: ProgressBar[];
+
+  private _panelConfig: PanelConfig.PanelConfig;
   private _element: any;
 
   private _seriesList: any;
 
   private _tooltip: GraphTooltip;
 
-  private statNameOptions = _.values(StatType);
+  private statNameOptions = _.values(PanelConfig.StatType);
   private statProgressTypeOptions = [ 'max value', 'shared' ];
   private coloringTypeOptions = [ 'auto', 'thresholds', 'key mapping' ];
-  private titleViewTypeOptions = _.values(TitleViewOptions);
+  private titleViewTypeOptions = _.values(PanelConfig.TitleViewOptions);
   private sortingOrderOptions = [ 'none', 'increasing', 'decreasing' ];
   private valueLabelTypeOptions = [ 'absolute', 'percentage' ];
   // TODO: change option names or add a tip in editor
   private mappingTypeOptions = ['datapoint to datapoint', 'target to datapoint'];
-  private tooltipModeOptions = _.values(TooltipMode);
+  private tooltipModeOptions = _.values(PanelConfig.TooltipMode);
 
   // field for updating tooltip on rendering and storing previous state
   private _lastHoverEvent: HoverEvent;
@@ -88,9 +60,9 @@ class Ctrl extends MetricsPanelCtrl {
   constructor($scope: any, $injector: any, public templateSrv: any) {
     super($scope, $injector);
 
-    _.defaults(this.panel, DEFAULTS);
+    _.defaults(this.panel, PanelConfig.DEFAULTS);
 
-    this._panelConfig = new PanelConfig(this.panel);
+    this._panelConfig = new PanelConfig.PanelConfig(this.panel);
     this._initStyles();
 
     this.mapper = new Mapper(this._panelConfig, this.templateSrv);
@@ -161,7 +133,7 @@ class Ctrl extends MetricsPanelCtrl {
       item.title, // previously wwe showed serie.alias || serie.target
       [{
         value: item.formattedValue,
-        color: item.color
+        color: item.colors[0] // TODO: use other colors here
       }]
     ));
     this._tooltip.show(event.event, tooltipItems, this.panel.tooltipMode);
