+++ src/panel_config.ts
@@ -1,3 +1,47 @@
+export enum StatType {
+  CURRENT = 'current',
+  MIN = 'min',
+  MAX = 'max',
+  TOTAL = 'total'
+};
+
+export enum TitleViewOptions {
+  SEPARATE_TITLE_LINE = 'Separate title line',
+  INLINE = 'Inline'
+};
+
+export enum TooltipMode {
+  NONE = 'none',
+  SINGLE = 'single',
+  ALL_SERIES = 'all series'
+};
+
+export const DEFAULTS = {
+  keyColumn: '',
+  // TODO: skip multiple columns
+  skipColumn: '',
+  statNameOptionValue: StatType.CURRENT,
+  statProgressType: 'shared',
+  statProgressMaxValue: null,
+  coloringType: 'auto',
+  titleViewType: TitleViewOptions.SEPARATE_TITLE_LINE,
+  sortingOrder: 'none',
+  valueLabelType: 'percentage',
+  mappingType: 'datapoint to datapoint',
+  alias: '',
+  prefix: '',
+  postfix: '',
+  thresholds: '10, 30',
+  // https://github.com/grafana/grafana/blob/v4.1.1/public/app/plugins/panel/singlestat/module.ts#L57
+  colors: ['rgba(245, 54, 54, 0.9)', 'rgba(237, 129, 40, 0.89)', 'rgba(50, 172, 45, 0.97)'],
+  colorsKeyMappingDefault: 'rgba(245, 54, 54, 0.9)',
+  colorKeyMappings: [],
+  nullMapping: undefined,
+  tooltipMode: TooltipMode.ALL_SERIES,
+  opacity: 0.5
+};
+
+
 export class PanelConfig {
   private _panel: any;
   public constructor(panel: any) {
